fix: menu navigation and session_start app context bugs
- open_main_menu: use AT-SPI Press action instead of coordinate
  click, fallback to click() for compatibility; select last
  DTitlebarDWindowOptionButton (index=-1) to handle multi-window
  apps like dde-file-manager
- session_start: stop overwriting context["app"] with launch command
  path in both AT and YAML executors, preserving AT-SPI registered
  name from suite/testcase config
- _get_dog: add os.path.basename() for path-style app names
- SuiteActionStep.wait: change type from int to float for finer
  wait control (e.g. 0.5s)

+++ src/at/parser/models.py
@@ -165,7 +165,7 @@ class SuiteActionStep(BaseModel):
     ref: Optional[str] = None
     selector: Optional[dict[str, Any]] = None
     command: Optional[str] = None
-    wait: Optional[int] = None
+    wait: Optional[float] = None
     do: Optional[str] = None
     x: Optional[int] = None
     y: Optional[int] = None
